backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
SQLAlchemy==2.0.25
email-validator==2.0.0.post2
bcrypt==4.0.1
